requirements.txt
Django==1.10.4
django-filter==1.0.1
djangorestframework==3.5.3
drf-nested-routers==0.11.1
Markdown==2.6.7
